+++ package/CHANGELOG.md
@@ -1,5 +1,17 @@
 # nexo-editor
 
+## 0.3.0
+
+### Minor Changes
+
+- reduced package size
+
+## 0.2.0
+
+### Minor Changes
+
+- reduced the package size
+
 ## 0.1.0
 
 ### Minor Changes